chore: support deleting multiple users
