--- apps/box/src/app/api/webhooks/elba/users/delete-user/service.ts
@@ -1,17 +1,0 @@
-import { inngest } from '@/inngest/client';
-
-export const deleteUsers = async ({
-  userId,
-  organisationId,
-}: {
-  userId: string;
-  organisationId: string;
-}) => {
-  await inngest.send({
-    name: 'box/users.delete.requested',
-    data: {
-      organisationId,
-      userId,
-    },
-  });
-};
