+++ apps/box/src/app/api/webhooks/elba/users/delete-users/route.ts
@@ -5,9 +5,12 @@ import { deleteUsers } from './service';
 export async function POST(request: NextRequest) {
   const data: unknown = await request.json();
 
-  const { id: userId, organisationId } = parseWebhookEventData('users.delete_user_requested', data);
+  const { ids: userIds, organisationId } = parseWebhookEventData(
+    'users.delete_users_requested',
+    data
+  );
 
-  await deleteUsers({ userId, organisationId });
+  await deleteUsers({ userIds, organisationId });
 
   return new NextResponse();
 }
