+++ apps/box/src/app/api/webhooks/elba/users/delete-users/service.ts
@@ -1,0 +1,19 @@
+import { inngest } from '@/inngest/client';
+
+export const deleteUsers = async ({
+  userIds,
+  organisationId,
+}: {
+  userIds: string[];
+  organisationId: string;
+}) => {
+  await inngest.send(
+    userIds.map((userId) => ({
+      name: 'box/users.delete.requested',
+      data: {
+        organisationId,
+        userId,
+      },
+    }))
+  );
+};
